# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ord",
    "description": "◉ Ordinal wallet and block explorer",
    "license": {
      "name": "CC0-1.0"
    },
    "version": "0.14.1"
  },
  "paths": {
...
    "/api/v1/ord/id/{id}/inscription": {
      "get": {
        "tags": [
          "ord"
        ],
        "summary": "Retrieve the inscription infomation with the specified inscription id.",
        "description": "Retrieve the inscription infomation with the specified inscription id.",
        "operationId": "ord_inscription_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "inscription ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain inscription infomation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOrdInscription"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ord/number/{number}/inscription": {
      "get": {
        "tags": [
          "ord"
        ],
        "summary": "Retrieve the inscription infomation with the specified inscription number.",
        "description": "Retrieve the inscription infomation with the specified inscription number.",
        "operationId": "ord_inscription_number",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "description": "inscription number",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain inscription infomation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOrdInscription"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string",
                "description": "Internal server error."
              }
            },
            "example": {
              "code": 1,
              "msg": "internal error"
            }
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "Bad request."
              }
            },
            "example": {
              "code": 1,
              "msg": "bad request"
            }
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "Resource not found."
              }
            },
            "example": {
              "code": 1,
              "msg": "not found"
            }
          }
        ]
      },
...
      "ApiOrdInscription": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/ord.ApiInscription"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "ScriptPubkey": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "address"
            ],
            "properties": {
              "address": {
                "type": "string",
                "description": "Address."
              }
            }
          },
          {
            "type": "object",
            "required": [
              "nonStandard"
            ],
            "properties": {
              "nonStandard": {
                "type": "string",
                "description": "Non-standard script hash."
              }
            }
          }
        ]
      },
...
      "ord.ApiContentEncoding": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "decode",
              "type"
            ],
            "properties": {
              "decode": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "br"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "unknown"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
      "ord.ApiInscription": {
        "type": "object",
        "required": [
          "id",
          "number",
          "contentEncoding",
          "genesisHeight",
          "genesisTimestamp",
          "location",
          "collections",
          "charms"
        ],
        "properties": {
          "charms": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Charms of Inscriptions."
          },
          "collections": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Collections of Inscriptions."
          },
          "content": {
            "type": "string",
            "description": "The inscription content body.",
            "nullable": true
          },
          "contentEncoding": {
            "$ref": "#/components/schemas/ord.ApiContentEncoding"
          },
          "contentLength": {
            "type": "integer",
            "description": "The inscription content body length.",
            "nullable": true,
            "minimum": 0
          },
          "contentType": {
            "type": "string",
            "description": "The inscription content type.",
            "nullable": true
          },
          "delegate": {
            "type": "string",
            "description": "The delegate inscription id of the inscription.",
            "nullable": true
          },
          "genesisHeight": {
            "type": "integer",
            "format": "uint32",
            "description": "The inscription genesis block height.",
            "minimum": 0
          },
          "genesisTimestamp": {
            "type": "integer",
            "format": "uint32",
            "description": "The inscription genesis timestamp.",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "description": "The inscription id."
          },
          "location": {
            "type": "string",
            "description": "The inscription location."
          },
          "metadata": {
            "type": "string",
            "description": "The inscription metadata.",
            "nullable": true
          },
          "metaprotocol": {
            "type": "string",
            "description": "The inscription metaprotocol.",
            "nullable": true
          },
          "number": {
            "type": "integer",
            "format": "int32",
            "description": "The inscription number."
          },
          "owner": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ScriptPubkey"
              }
            ],
            "nullable": true
          },
          "parent": {
            "type": "string",
            "description": "The inscription parent inscription id.",
            "nullable": true
          },
          "pointer": {
            "type": "integer",
            "format": "int64",
            "description": "The inscription pointer.",
            "nullable": true,
            "minimum": 0
          },
          "rune": {
            "type": "string",
            "description": "Rune tag of the inscription.",
            "nullable": true
          },
          "sat": {
            "type": "integer",
            "format": "int64",
            "description": "The inscription sat index.",
            "nullable": true,
            "minimum": 0
          }
        }
      },
...
    }
  }
}